refactor(projectSubmissions): seperate out filter out
