+++ src/frontend/src/components/ProjectSubmissions/Table/index.tsx
@@ -1,31 +1,18 @@
 import React, { useEffect, useState } from 'react';
-import { Link, useNavigate, useSearchParams } from 'react-router-dom';
+import { Link, useSearchParams } from 'react-router-dom';
 import { format } from 'date-fns';
 import { Tooltip } from '@mui/material';
-import { Loader2 } from 'lucide-react';
-
 import AssetModules from '@/shared/AssetModules.js';
 import CoreModules from '@/shared/CoreModules.js';
-import windowDimention from '@/hooks/WindowDimension';
-
-import Button from '@/components/common/Button';
 import { Modal } from '@/components/common/Modal';
-import { CustomSelect } from '@/components/common/Select.js';
-import DateRangePicker from '@/components/common/DateRangePicker';
 import Table, { TableHeader } from '@/components/common/CustomTable';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@/components/common/Dropdown';
 import UpdateReviewStatusModal from '@/components/ProjectSubmissions/Table/UpdateReviewStatusModal';
-import { reviewStateData } from '@/constants/projectSubmissionsConstants';
-
+import Filter from '@/components/ProjectSubmissions/Table/Filter';
 import { useAppDispatch, useAppSelector } from '@/types/reduxTypes';
-import { task_state, task_event, project_status } from '@/types/enums';
+import { project_status } from '@/types/enums';
 import { filterType } from '@/store/types/ISubmissions';
 import { SubmissionActions } from '@/store/slices/SubmissionSlice';
-
-import { CreateTaskEvent } from '@/api/TaskEvent';
-import { ConvertXMLToJOSM, DownloadProjectSubmission } from '@/api/task';
 import { SubmissionFormFieldsService, SubmissionTableService } from '@/api/SubmissionService';
-
 import filterParams from '@/utilfunctions/filterParams';
 import { camelToFlat } from '@/utilfunctions/commonUtils';
 import useDocumentTitle from '@/utilfunctions/useDocumentTitle';
@@ -46,31 +33,19 @@ const SubmissionsTable = ({ toggleView }) => {
   };
   const [filter, setFilter] = useState<filterType>(initialFilterState);
 
-  const { windowSize } = windowDimention();
   const dispatch = useAppDispatch();
   const params = CoreModules.useParams();
-  const navigate = useNavigate();
 
   const projectId = params.projectId;
   const submissionFormFields = useAppSelector((state) => state.submission.submissionFormFields);
   const submissionTableData = useAppSelector((state) => state.submission.submissionTableData);
   const submissionFormFieldsLoading = useAppSelector((state) => state.submission.submissionFormFieldsLoading);
   const submissionTableDataLoading = useAppSelector((state) => state.submission.submissionTableDataLoading);
-  const submissionTableRefreshing = useAppSelector((state) => state.submission.submissionTableRefreshing);
-  const taskInfo = useAppSelector((state) => state.task.taskInfo);
   const projectInfo = useAppSelector((state) => state.project.projectInfo);
   const josmEditorError = useAppSelector((state) => state.task.josmEditorError);
-  const downloadSubmissionLoading = useAppSelector((state) => state.task.downloadSubmissionLoading);
-  const authDetails = CoreModules.useAppSelector((state) => state.login.authDetails);
-  const updateTaskStatusLoading = useAppSelector((state) => state.common.loading);
 
   const projectData = useAppSelector((state) => state.project.projectTaskBoundries);
   const projectIndex = projectData.findIndex((project) => project.id == +projectId);
-  const currentStatus = {
-    ...projectData?.[projectIndex]?.taskBoundries?.filter((task) => {
-      return filter.task_id && task?.id === +filter.task_id;
-    })?.[0],
-  };
   const taskList = projectData[projectIndex]?.taskBoundries;
 
   const isProjectManager = useIsProjectManager(projectId as string);
@@ -86,24 +61,6 @@ const SubmissionsTable = ({ toggleView }) => {
       ? new Date(initialFilterState.submitted_date_range.split(',')[1])
       : null,
   });
-
-  const submissionDownloadTypes: { type: 'csv' | 'json' | 'geojson'; label: string; loading: boolean }[] = [
-    {
-      type: 'csv',
-      label: 'Download as Csv',
-      loading: downloadSubmissionLoading.fileType === 'csv' && downloadSubmissionLoading.loading,
-    },
-    {
-      type: 'json',
-      label: 'Download as Json',
-      loading: downloadSubmissionLoading.fileType === 'json' && downloadSubmissionLoading.loading,
-    },
-    {
-      type: 'geojson',
-      label: 'Download as GeoJson',
-      loading: downloadSubmissionLoading.fileType === 'geojson' && downloadSubmissionLoading.loading,
-    },
-  ];
 
   useEffect(() => {
     if (!dateRange.start || !dateRange.end) return;
@@ -209,34 +166,6 @@ const SubmissionsTable = ({ toggleView }) => {
     return value ? (typeof value === 'object' ? '-' : value) : '-';
   }
 
-  // const uploadToJOSM = () => {
-  //   dispatch(ConvertXMLToJOSM(`${VITE_API_URL}/submission/get_osm_xml/${projectId}`, projectInfo?.outline?.bbox));
-  // };
-
-  const handleDownload = (downloadType: 'csv' | 'json' | 'geojson') => {
-    dispatch(
-      DownloadProjectSubmission(`${VITE_API_URL}/submission/download`, projectInfo.name!, {
-        project_id: projectId,
-        submitted_date_range: filter?.submitted_date_range,
-        file_type: downloadType,
-      }),
-    );
-  };
-
-  const handleTaskMap = async () => {
-    await dispatch(
-      CreateTaskEvent(
-        `${VITE_API_URL}/tasks/${currentStatus.id}/event`,
-        task_event.GOOD,
-        projectId,
-        filter?.task_id || '',
-        authDetails || {},
-        { project_id: projectId },
-      ),
-    );
-    navigate(`/project/${projectId}`);
-  };
-
   useEffect(() => {
     const filteredParams = filterParams(filter);
     setSearchParams({ tab: 'table', ...filteredParams });
@@ -260,143 +189,17 @@ const SubmissionsTable = ({ toggleView }) => {
         }}
       />
       <UpdateReviewStatusModal />
-      <div className="fmtm-flex xl:fmtm-items-end xl:fmtm-justify-between fmtm-flex-col lg:fmtm-flex-row fmtm-gap-4 fmtm-mb-6">
-        <div
-          className={`${
-            windowSize.width < 2000 ? 'fmtm-w-full md:fmtm-w-fit' : 'fmtm-w-fit'
-          } fmtm-flex xl:fmtm-items-end fmtm-gap-2 xl:fmtm-gap-4 fmtm-rounded-lg fmtm-flex-col sm:fmtm-flex-row fmtm-order-2 md:-fmtm-order-1`}
-        >
-          <DropdownMenu modal={false}>
-            <DropdownMenuTrigger>
-              <Button variant="secondary-red">
-                <AssetModules.TuneIcon style={{ fontSize: '20px' }} />
-                FILTER{' '}
-                <div className="fmtm-bg-primaryRed fmtm-text-white fmtm-rounded-full fmtm-w-4 fmtm-h-4 fmtm-flex fmtm-justify-center fmtm-items-center">
-                  <p>{Object.values(filter).filter((filterObjValue) => filterObjValue).length}</p>
-                </div>
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent className="fmtm-z-[50]" align="start">
-              <div
-                className={`fmtm-w-fit fmtm-max-w-[90vw] -fmtm-bottom-20 fmtm-bg-white fmtm-px-4 fmtm-rounded-lg fmtm-shadow-2xl fmtm-pb-4 fmtm-pt-2 fmtm-gap-4 fmtm-items-end fmtm-flex fmtm-flex-wrap`}
-              >
-                <div className={`${windowSize.width < 500 ? 'fmtm-w-full' : 'fmtm-w-[11rem]'}`}>
-                  <CustomSelect
-                    title="Task Id"
-                    placeholder="Select"
-                    data={taskInfo}
-                    dataKey="value"
-                    value={filter?.task_id?.toString() || ''}
-                    valueKey="task_id"
-                    label="task_id"
-                    onValueChange={(value) => value && setFilter((prev) => ({ ...prev, task_id: value.toString() }))}
-                    className="fmtm-text-grey-700 fmtm-text-sm !fmtm-mb-0 fmtm-bg-white"
-                  />
-                </div>
-                <div className={`${windowSize.width < 500 ? 'fmtm-w-full' : 'fmtm-w-[11rem]'}`}>
-                  <CustomSelect
-                    title="Review State"
-                    placeholder="Select"
-                    data={reviewStateData}
-                    dataKey="value"
-                    value={filter?.review_state || ''}
-                    valueKey="value"
-                    label="label"
-                    onValueChange={(value) =>
-                      value && setFilter((prev) => ({ ...prev, review_state: value.toString() }))
-                    }
-                    errorMsg=""
-                    className="fmtm-text-grey-700 fmtm-text-sm !fmtm-mb-0 fmtm-bg-white"
-                  />
-                </div>
-                <div className={`${windowSize.width < 500 ? 'fmtm-w-full' : 'fmtm-w-[12rem]'}`}>
-                  <DateRangePicker
-                    title="Submitted Date"
-                    startDate={dateRange?.start}
-                    endDate={dateRange?.end}
-                    setStartDate={(date) => setDateRange((prev) => ({ ...prev, start: date }))}
-                    setEndDate={(date) => setDateRange((prev) => ({ ...prev, end: date }))}
-                    className="fmtm-text-grey-700 fmtm-text-sm !fmtm-mb-0 fmtm-w-full"
-                  />
-                </div>
-                <div className={`${windowSize.width < 500 ? 'fmtm-w-full' : 'fmtm-w-[11rem]'}`}>
-                  <p className={`fmtm-text-grey-700 fmtm-text-sm fmtm-font-semibold !fmtm-bg-transparent`}>
-                    Submitted By
-                  </p>
-                  <div className="fmtm-border fmtm-border-gray-300 sm:fmtm-w-fit fmtm-flex fmtm-bg-white fmtm-items-center fmtm-px-1">
-                    <input
-                      type="search"
-                      className="fmtm-h-[1.9rem] fmtm-p-2 fmtm-w-full fmtm-outline-none"
-                      placeholder="Search User"
-                      onChange={(e) => {
-                        setSubmittedBy(e.target.value);
-                      }}
-                    ></input>
-                    <AssetModules.SearchIcon className="fmtm-text-[#9B9999] fmtm-cursor-pointer" />
-                  </div>
-                </div>
-                <Button
-                  variant="secondary-red"
-                  onClick={clearFilters}
-                  disabled={submissionTableDataLoading || submissionFormFieldsLoading}
-                >
-                  Reset Filter
-                </Button>
-              </div>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          <div className="fmtm-flex fmtm-gap-4">
-            {/* comment out button until JOSM is implemented */}
-            {/* <Button variant="primary-red" onClick={uploadToJOSM}>
-              <AssetModules.FileDownloadIcon className="!fmtm-text-xl" />
-              UPLOAD TO JOSM
-            </Button> */}
-            <DropdownMenu>
-              <DropdownMenuTrigger>
-                <Button variant="link-grey">
-                  <AssetModules.FileDownloadIcon className="!fmtm-text-xl" />
-                  DOWNLOAD
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="start" className="fmtm-z-[5000] fmtm-bg-white">
-                {submissionDownloadTypes?.map((submissionDownload) => (
-                  <DropdownMenuItem
-                    key={submissionDownload.type}
-                    disabled={submissionDownload.loading}
-                    onSelect={() => handleDownload(submissionDownload.type)}
-                  >
-                    <div className="fmtm-flex fmtm-gap-2 fmtm-items-center">
-                      <p className="fmtm-text-base">{submissionDownload.label}</p>
-                      {submissionDownload.loading && (
-                        <Loader2 className="fmtm-h-4 fmtm-w-4 fmtm-animate-spin fmtm-text-primaryRed" />
-                      )}
-                    </div>
-                  </DropdownMenuItem>
-                ))}
-              </DropdownMenuContent>
-            </DropdownMenu>
-            <Button
-              variant="link-grey"
-              onClick={refreshTable}
-              isLoading={(submissionTableDataLoading || submissionFormFieldsLoading) && submissionTableRefreshing}
-              disabled={submissionTableDataLoading || submissionFormFieldsLoading}
-            >
-              <AssetModules.RefreshIcon className="!fmtm-text-xl" />
-              REFRESH
-            </Button>
-          </div>
-        </div>
-        <div className="fmtm-w-full fmtm-flex fmtm-items-center fmtm-justify-end xl:fmtm-w-fit fmtm-gap-3">
-          {filter?.task_id &&
-            projectData?.[projectIndex]?.taskBoundries?.find((task) => task?.id === +filter?.task_id)?.task_state ===
-              task_state.LOCKED_FOR_VALIDATION && (
-              <Button variant="primary-red" onClick={handleTaskMap} isLoading={updateTaskStatusLoading}>
-                MARK AS VALIDATED
-              </Button>
-            )}
-          {toggleView}
-        </div>
-      </div>
+      <Filter
+        toggleView={toggleView}
+        filter={filter}
+        setFilter={setFilter}
+        dateRange={dateRange}
+        setDateRange={setDateRange}
+        submittedBy={submittedBy}
+        setSubmittedBy={setSubmittedBy}
+        clearFilters={clearFilters}
+        refreshTable={refreshTable}
+      />
       {submissionTableDataLoading || submissionFormFieldsLoading ? (
         <SubmissionsTableSkeleton />
       ) : (
